Move instance/physical_device.rs to device/physical.rs, rewrite some of it
