+++ vulkano/src/device/properties.rs
@@ -1,9 +1,9 @@
 use crate::descriptor::descriptor::ShaderStages;
-use crate::image::{SampleCount, SampleCounts};
-use crate::instance::{
+use crate::device::physical::{
     ConformanceVersion, DriverId, PhysicalDeviceType, PointClippingBehavior, ShaderCoreProperties,
     ShaderFloatControlsIndependence, SubgroupFeatures,
 };
+use crate::image::{SampleCount, SampleCounts};
 use crate::render_pass::ResolveModes;
 use crate::Version;
 use std::convert::TryInto;
