+++ vulkano/src/image/immutable.rs
@@ -16,6 +16,7 @@ use crate::command_buffer::CommandBufferExecFuture;
 use crate::command_buffer::CommandBufferUsage;
 use crate::command_buffer::PrimaryAutoCommandBuffer;
 use crate::command_buffer::PrimaryCommandBuffer;
+use crate::device::physical::QueueFamily;
 use crate::device::Device;
 use crate::device::Queue;
 use crate::format::Format;
@@ -32,7 +33,6 @@ use crate::image::ImageLayout;
 use crate::image::ImageUsage;
 use crate::image::MipmapsCount;
 use crate::image::SampleCount;
-use crate::instance::QueueFamily;
 use crate::memory::pool::AllocFromRequirementsFilter;
 use crate::memory::pool::AllocLayout;
 use crate::memory::pool::MappingRequirement;
