+++ vulkano/src/instance/mod.rs
@@ -31,7 +31,7 @@
 //! # use vulkano::instance::Instance;
 //! # use vulkano::instance::InstanceExtensions;
 //! # use vulkano::Version;
-//! use vulkano::instance::PhysicalDevice;
+//! use vulkano::device::physical::PhysicalDevice;
 //!
 //! # let instance = Instance::new(None, Version::V1_1, &InstanceExtensions::none(), None).unwrap();
 //! for physical_device in PhysicalDevice::enumerate(&instance) {
@@ -60,21 +60,6 @@ pub use self::layers::LayerProperties;
 pub use self::layers::LayersIterator;
 pub use self::layers::LayersListError;
 pub use self::loader::LoadingError;
-pub use self::physical_device::ConformanceVersion;
-pub use self::physical_device::DriverId;
-pub use self::physical_device::MemoryHeap;
-pub use self::physical_device::MemoryHeapsIter;
-pub use self::physical_device::MemoryType;
-pub use self::physical_device::MemoryTypesIter;
-pub use self::physical_device::PhysicalDevice;
-pub use self::physical_device::PhysicalDeviceType;
-pub use self::physical_device::PhysicalDevicesIter;
-pub use self::physical_device::PointClippingBehavior;
-pub use self::physical_device::QueueFamiliesIter;
-pub use self::physical_device::QueueFamily;
-pub use self::physical_device::ShaderCoreProperties;
-pub use self::physical_device::ShaderFloatControlsIndependence;
-pub use self::physical_device::SubgroupFeatures;
 pub use crate::extensions::{
     ExtensionRestriction, ExtensionRestrictionError, SupportedExtensionsError,
 };
@@ -85,4 +70,3 @@ pub(crate) mod extensions;
 mod instance;
 mod layers;
 pub mod loader;
-mod physical_device;
