+++ vulkano/src/memory/device_memory.rs
@@ -8,9 +8,9 @@
 // according to those terms.
 
 use crate::check_errors;
+use crate::device::physical::MemoryType;
 use crate::device::Device;
 use crate::device::DeviceOwned;
-use crate::instance::MemoryType;
 use crate::memory::Content;
 use crate::memory::DedicatedAlloc;
 use crate::memory::ExternalMemoryHandleType;
@@ -269,7 +269,7 @@ impl<'a> DeviceMemoryBuilder<'a> {
 
         let mut export_handle_bits = ash::vk::ExternalMemoryHandleTypeFlags::empty();
         if self.dedicated_info.is_some() {
-            if !self.device.loaded_extensions().khr_dedicated_allocation {
+            if !self.device.enabled_extensions().khr_dedicated_allocation {
                 return Err(DeviceMemoryAllocError::MissingExtension(
                     "khr_dedicated_allocation",
                 ));
@@ -291,7 +291,7 @@ impl<'a> DeviceMemoryBuilder<'a> {
             if !(export_handle_bits & ash::vk::ExternalMemoryHandleTypeFlags::DMA_BUF_EXT)
                 .is_empty()
             {
-                if !self.device.loaded_extensions().ext_external_memory_dma_buf {
+                if !self.device.enabled_extensions().ext_external_memory_dma_buf {
                     return Err(DeviceMemoryAllocError::MissingExtension(
                         "ext_external_memory_dmabuf",
                     ));
@@ -300,7 +300,7 @@ impl<'a> DeviceMemoryBuilder<'a> {
 
             if !(export_handle_bits & ash::vk::ExternalMemoryHandleTypeFlags::OPAQUE_FD).is_empty()
             {
-                if !self.device.loaded_extensions().khr_external_memory_fd {
+                if !self.device.enabled_extensions().khr_external_memory_fd {
                     return Err(DeviceMemoryAllocError::MissingExtension(
                         "khr_external_memory_fd",
                     ));
@@ -310,7 +310,7 @@ impl<'a> DeviceMemoryBuilder<'a> {
             if !(import_handle_bits & ash::vk::ExternalMemoryHandleTypeFlags::DMA_BUF_EXT)
                 .is_empty()
             {
-                if !self.device.loaded_extensions().ext_external_memory_dma_buf {
+                if !self.device.enabled_extensions().ext_external_memory_dma_buf {
                     return Err(DeviceMemoryAllocError::MissingExtension(
                         "ext_external_memory_dmabuf",
                     ));
@@ -319,7 +319,7 @@ impl<'a> DeviceMemoryBuilder<'a> {
 
             if !(import_handle_bits & ash::vk::ExternalMemoryHandleTypeFlags::OPAQUE_FD).is_empty()
             {
-                if !self.device.loaded_extensions().khr_external_memory_fd {
+                if !self.device.enabled_extensions().khr_external_memory_fd {
                     return Err(DeviceMemoryAllocError::MissingExtension(
                         "khr_external_memory_fd",
                     ));
